Moved to center_scrap, refactord center_iterator and moved dedup to "all" iterator
