+++ scraper/doctolib/doctolib_center_scrap.py
@@ -1,3 +1,10 @@
+from scraper.pattern.center_location import CenterLocation
+from scraper.pattern.scraper_result import DRUG_STORE, GENERAL_PRACTITIONER, VACCINATION_CENTER
+from typing import List
+from scraper.pattern.center_info import CenterInfo
+import requests
+
+
 import csv
 import io
 import json
@@ -10,7 +17,7 @@ from scraper.doctolib.doctolib_filters import parse_practitioner_type
 from utils.vmd_utils import departementUtils
 from utils.vmd_logger import enable_logger_for_production
 
-BASE_URL = 'https://www.doctolib.fr/vaccination-covid-19/france?page={0}'
+BASE_URL = 'http://www.doctolib.fr/vaccination-covid-19/france.json?page={0}'
 BOOKING_URL = 'https://www.doctolib.fr/booking/{0}.json'
 
 CENTER_TYPES = [
@@ -29,47 +36,117 @@ DOCTOLIB_DOMAINS = [
 ]
 
 
-def get_csv():
-    url = "https://www.data.gouv.fr/fr/datasets/r/5cb21a85-b0b0-4a65-a249-806a040ec372"
-    response = requests.get(url)
-    response.raise_for_status()
+def parse_doctolib_centers() -> List[dict]:
+    page_id = 1
+    page_has_centers = True
 
-    reader = io.StringIO(response.content.decode('utf8'))
-    csvreader = csv.DictReader(reader, delimiter=';')
-    for row in csvreader:
-        yield row
+    centers = []
 
+    while page_has_centers and page_id == 1:
+        print(f"Parsing page {page_id}")
+        centers_page = parse_page_centers(page_id)
+        centers += centers_page
 
-def get_doctolib_csv_urls():
-    urls = []
-    for center in get_csv():
-        url = center['rdv_site_web']
-        if len(url) == 0:
-            continue
-        is_doctolib_url = False
-        for domain in DOCTOLIB_DOMAINS:
-            if url.startswith(domain):
-                is_doctolib_url = True
-        if not is_doctolib_url:
-            continue
-        uri_info = parse.urlsplit(url)
-        urls.append(uri_info.path)
-    return urls
+        page_id += 1
+
+        if len(centers_page) == 0:
+            page_has_centers = False
+        print([center["nom"] for center in centers_page])
+
+    return centers
 
 
-def is_url_in_csv(url, list):
-    uri_info = parse.urlsplit(url)
-    path = uri_info.path
+def parse_page_centers(page_id) -> List[dict]:
+    r = requests.get(BASE_URL.format(page_id))
+    data = r.json()
 
-    return path in list
+    centers_page = [center_from_doctor_dict(payload)
+                    for payload in data["data"]["doctors"]]
+    return centers_page
 
 
-def parse_doctolib_business_hours(url_data, place):
+def center_from_doctor_dict(doctor_dict) -> dict:
+    nom = doctor_dict['name_with_title']
+    sub_addresse = doctor_dict["address"]
+    ville = doctor_dict["city"]
+    code_postal = doctor_dict["zipcode"]
+    addresse = f"{sub_addresse}, {code_postal} {ville}"
+    if doctor_dict['place_id']:
+        url_path = doctor_dict['link']+"?pid="+str(doctor_dict["place_id"])
+    else:
+        url_path = doctor_dict['link']
+    rdv_site_web = f"https://partners.doctolib.fr{url_path}"
+    type = center_type(url_path, nom)
+    dict_infos_center_page = get_dict_infos_center_page(url_path)
+    dict_infos_browse_page = {
+        "nom": nom,
+        "rdv_site_web": rdv_site_web,
+        "ville": ville,
+        "address": addresse,
+        "long_coor1": float(doctor_dict["position"]["lng"]),
+        "lat_coor1": float(doctor_dict["position"]["lat"]),
+        "type": type,
+        "com_insee": departementUtils.cp_to_insee(code_postal)
+    }
+    return {**dict_infos_center_page, **dict_infos_browse_page}
+
+
+def get_dict_infos_center_page(url_path: str) -> dict:
+    internal_api_url = BOOKING_URL.format(
+        parse.urlsplit(url_path).path.split("/")[-1])
+    data = requests.get(internal_api_url)
+    data.raise_for_status()
+    output = data.json().get('data', {})
+
+    # Parse place
+    places = output.get('places', {})
+    if places:
+        place = find_place(places, url_path)
+
+        # Parse place location
+        infos_page = {}
+        infos_page['address'] = place['full_address']
+        infos_page['long_coor1'] = place['longitude']
+        infos_page['lat_coor1'] = place['latitude']
+        infos_page["com_insee"] = departementUtils.cp_to_insee(
+            place["zipcode"])
+
+        # Parse landline number
+        if place.get('landline_number'):
+            infos_page['phone_number'] = place.get('landline_number', None)
+        if place.get('phone_number'):
+            infos_page['phone_number'] = place.get('phone_number', None)
+
+        infos_page["business_hours"] = parse_doctolib_business_hours(place)
+        return infos_page
+    else:
+        return {}
+
+
+def find_place(places, url_path):
+    pid = get_pid(url_path)
+    if pid:
+        for place in places:
+            if place["id"] == pid:
+                return place
+    return places[0]
+
+
+def get_pid(url_path) -> str:
+    split = url_path.split("?pid=")
+    if len(split) == 1:
+        return ""
+    else:
+        return split[1]
+
+
+def parse_doctolib_business_hours(place) -> dict:
     # Opening hours
     business_hours = dict()
-    keys = ["lundi", "mardi", "mercredi", "jeudi", "vendredi", "samedi", "dimanche"]
+    keys = ["lundi", "mardi", "mercredi",
+            "jeudi", "vendredi", "samedi", "dimanche"]
     if not place['opening_hours']:
-        return url_data
+        return None
 
     for opening_hour in place['opening_hours']:
         format_hours = ''
@@ -83,123 +160,18 @@ def parse_doctolib_business_hours(url_data, place):
             format_hours += f'{range[0]}-{range[1]}'
         business_hours[key_name] = format_hours
 
-    if business_hours:
-        url_data['business_hours'] = business_hours
-    return url_data
+    return business_hours
 
 
-def get_doctolib_center_data(url_data):
-    data = requests.get(BOOKING_URL.format(url_data['internal_api_name']))
-    data.raise_for_status()
-    output = data.json().get('data', {})
-
-    url_data['gid'] = f'd{output.get("profile", {}).get("id", 0)}'[:8]
-
-    # Parse place
-    place = output.get('places', {})
-
-    # Parse practitioner type
-    url_data['type'] = parse_practitioner_type(url_data['nom'], output)
-
-    if not place:
-        return url_data
-    place = place[0]  # for the moment
-
-    # Parse place location
-    url_data['address'] = place['full_address']
-    url_data['long_coor1'] = place['longitude']
-    url_data['lat_coor1'] = place['latitude']
-    url_data["com_insee"] = departementUtils.cp_to_insee(place["zipcode"])
-
-    # Parse landline number
-    if place.get('landline_number', None):
-        url_data['phone_number'] = place.get('landline_number', None)
-
-    url_data = parse_doctolib_business_hours(url_data, place)
-    return url_data
-
-
-def is_revelant_url(url):
-    href = url.get('href')
-    is_center = False
-    if len(href.split('/')) < 3:
-        return False
-
-    for center_type in CENTER_TYPES:
-        if href.startswith(f'/{center_type}/'):
-            is_center = True
-            break
-
-    if not is_center or not url.contents or not url.contents[0]:
-        return False
-    if not url.div or url.img:
-        return False
-    return True
-
-
-def check_duplicate(href):
-    if not href:
-        return None
-    return href.split('?')[0]
-
-
-def scrape_page(page_id, url_list, gids):
-    data = requests.get(BASE_URL.format(page_id))
-    data.raise_for_status()
-    output = data.text
-    soup = BeautifulSoup(output, features="html.parser")
-    center_urls = []
-    total_urls = 0  # Total url count in this page
-    # It even counts filtered URLs
-
-    for link in soup.findAll("a"):
-        href = link.get('href')
-        if not is_revelant_url(link):
-            continue
-        vp = str(link.div.string)
-        total_urls += 1
-        api_name = href.split('/')[-1]
-        if is_url_in_csv(href, url_list):
-            continue
-        center_data = {'rdv_site_web': href, 'nom': vp, 'internal_api_name': api_name}
-        center_data = get_doctolib_center_data(center_data)
-        if not center_data:
-            continue
-        if center_data['gid'] in gids:
-            continue
-        center_data['rdv_site_web'] = f'https://partners.doctolib.fr{href}'
-        gids.append(center_data['gid'])
-        center_urls.append(center_data)
-    return center_urls, total_urls
-
-
-def main():
-    logger = enable_logger_for_production()
-    center_urls = []
-    gids = []
-    url_list = get_doctolib_csv_urls()
-    i = 1
-
-    while i < 2000:
-        try:
-            centr, url_count = scrape_page(i, url_list, gids)
-            if url_count == 0:
-                logger.info("Page: {0} <-> No center on this page. Stopping.".format(i))
-                break
-            center_urls.extend(centr)
-            logger.info(
-                "Page: {0} <-> Found {1} centers not in data.gouv CSV.".format(i, len(center_urls)))
-        except Exception as e:
-            logger.warning("Unable to scrape Doctolib page {0}".format(i))
-            break
-        i += 1
-    if len(center_urls) == 0:
-        logger.error("No Doctolib center found. Banned?")
-        exit(1)
-    file = open('data/output/doctolib-centers.json', 'w')
-    file.write(json.dumps(center_urls, indent=2))
-    file.close()
+def center_type(url_path: str, nom: str) -> str:
+    if "pharmacie" in nom.lower() or "pharmacie" in url_path:
+        return DRUG_STORE
+    if "medecin" in url_path or "medecin" in nom.lower():
+        return GENERAL_PRACTITIONER
+    return VACCINATION_CENTER
 
 
 if __name__ == "__main__":
-    main()
+    centers = parse_doctolib_centers()
+    with open('data/output/doctolib-centers.json', 'w') as f:
+        f.write(json.dumps(centers, indent=2))
