+++ scraper/scraper.py
@@ -19,6 +19,7 @@ from scraper.pattern.scraper_result import ScraperResult, VACCINATION_CENTER
 from utils.vmd_logger import enable_logger_for_production, enable_logger_for_debug
 from utils.vmd_utils import departementUtils
 from .doctolib.doctolib import fetch_slots as doctolib_fetch_slots
+from .doctolib.doctolib import center_iterator as doctolib_center_iterator
 from .keldoc.keldoc import fetch_slots as keldoc_fetch_slots
 from .maiia import fetch_slots as maiia_fetch_slots
 from .ordoclic import centre_iterator as ordoclic_centre_iterator
@@ -52,7 +53,9 @@ def scrape_debug(urls):
             result = fetch_centre_slots(rdv_site_web, start_date)
         except Exception as e:
             logger.exception(f"erreur lors du traitement")
-        logger.info(f'{result.platform!s:16} {result.next_availability or ""!s:32}')
+        logger.info(
+            f'{result.platform!s:16} {result.next_availability or ""!s:32}')
+
 
 def scrape():
     with Pool(POOL_SIZE) as pool:
@@ -61,7 +64,8 @@ def scrape():
             centre_iterator(),
             1
         )
-        compte_centres, compte_centres_avec_dispo, compte_bloqués = export_data(centres_cherchés)
+        compte_centres, compte_centres_avec_dispo, compte_bloqués = export_data(
+            centres_cherchés)
         logger.info(
             f"{compte_centres_avec_dispo} centres de vaccination avaient des disponibilités sur {compte_centres} scannés")
         if compte_centres_avec_dispo == 0:
@@ -84,10 +88,12 @@ def cherche_prochain_rdv_dans_centre(centre):
         result = fetch_centre_slots(centre['rdv_site_web'], start_date)
         center_data.fill_result(result)
     except ScrapeError as scrape_error:
-        logger.error(f"erreur lors du traitement de la ligne avec le gid {centre['gid']} {str(scrape_error)}")
+        logger.error(
+            f"erreur lors du traitement de la ligne avec le gid {centre['gid']} {str(scrape_error)}")
         has_error = scrape_error
     except Exception as e:
-        logger.error(f"erreur lors du traitement de la ligne avec le gid {centre['gid']}")
+        logger.error(
+            f"erreur lors du traitement de la ligne avec le gid {centre['gid']}")
         traceback.print_exc()
 
     if has_error is None:
@@ -114,7 +120,8 @@ def fix_scrap_urls(url):
 
     # Fix Keldoc
     if url.startswith("https://www.keldoc.com/"):
-        url = url.replace("https://www.keldoc.com/", "https://vaccination-covid.keldoc.com/")
+        url = url.replace("https://www.keldoc.com/",
+                          "https://vaccination-covid.keldoc.com/")
     # Clean Doctolib
     if url.startswith('https://partners.doctolib.fr') or url.startswith('https://www.doctolib.fr'):
         u = urlparse(url)
@@ -156,12 +163,14 @@ def export_data(centres_cherchés, outpath_format='data/output/{}.json'):
     try:
         with open(outpath, "r") as fichier:
             centres_fermes_gouv = json.load(fichier)
-            centres_fermes_gouv = centres_fermes_gouv["centres_fermes"].values()
+            centres_fermes_gouv = centres_fermes_gouv["centres_fermes"].values(
+            )
     except Exception as e:
         centres_fermes_gouv = {}
         logger.warning("Pas de fichier centres_non_pris_en_compte_gouv.json")
 
-    centres_fermes_dispo_gouv = {"centres_disponibles": [],"centres_indisponibles":[]}
+    centres_fermes_dispo_gouv = {
+        "centres_disponibles": [], "centres_indisponibles": []}
 
     internal_ids = []
     for centre in centres_cherchés:
@@ -179,25 +188,30 @@ def export_data(centres_cherchés, outpath_format='data/output/{}.json'):
             if code_departement in par_departement:
                 erreur = centre.pop('erreur', None)
                 if centre['prochain_rdv'] is None:
-                    par_departement[code_departement]['centres_indisponibles'].append(centre)
+                    par_departement[code_departement]['centres_indisponibles'].append(
+                        centre)
                     if isinstance(erreur, BlockedByDoctolibError):
                         par_departement[code_departement]['doctolib_bloqué'] = True
                         bloqués_doctolib += 1
                 else:
                     compte_centres_avec_dispo += 1
-                    par_departement[code_departement]['centres_disponibles'].append(centre)
+                    par_departement[code_departement]['centres_disponibles'].append(
+                        centre)
             else:
                 logger.warning(
                     f"le centre {centre['nom']} ({code_departement}) n'a pas pu être rattaché à un département connu")
         else:
             if centre['prochain_rdv'] is None:
-                centres_fermes_dispo_gouv["centres_indisponibles"].append(centre)
+                centres_fermes_dispo_gouv["centres_indisponibles"].append(
+                    centre)
             else:
-                logger.info(f"le centre {centre['nom']} est fermé et a un rdv disponible")
+                logger.info(
+                    f"le centre {centre['nom']} est fermé et a un rdv disponible")
                 centres_fermes_dispo_gouv["centres_disponibles"].append(centre)
-    
+
     nb_fermes_dispos = len(centres_fermes_dispo_gouv["centres_disponibles"])
-    logger.info(f" {nb_fermes_dispos} centres sont fermés et ont un rdv disponible")
+    logger.info(
+        f" {nb_fermes_dispos} centres sont fermés et ont un rdv disponible")
 
     outpath = outpath_format.format("centres_fermes_dispo_gouv")
     with open(outpath, "w") as fichier:
@@ -208,9 +222,11 @@ def export_data(centres_cherchés, outpath_format='data/output/{}.json'):
         json.dump(par_departement, info_centres, indent=2)
 
     for code_departement, disponibilités in par_departement.items():
-        disponibilités['last_updated'] = dt.datetime.now(tz=pytz.timezone('Europe/Paris')).isoformat()
+        disponibilités['last_updated'] = dt.datetime.now(
+            tz=pytz.timezone('Europe/Paris')).isoformat()
         if 'centres_disponibles' in disponibilités:
-            disponibilités['centres_disponibles'] = sorted(disponibilités['centres_disponibles'], key=sort_center)
+            disponibilités['centres_disponibles'] = sorted(
+                disponibilités['centres_disponibles'], key=sort_center)
         outpath = outpath_format.format(code_departement)
         logger.debug(f'writing result to {outpath} file')
         with open(outpath, "w") as outfile:
@@ -250,7 +266,8 @@ def fetch_centre_slots(rdv_site_web, start_date, fetch_map: dict = None):
     platform = None
     for scraper_name in fetch_map:
         scraper = fetch_map[scraper_name]
-        scrap = sum([1 if rdv_site_web.startswith(url) else 0 for url in scraper.get('urls', [])])
+        scrap = sum([1 if rdv_site_web.startswith(url)
+                    else 0 for url in scraper.get('urls', [])])
         if scrap == 0:
             continue
         platform = scraper_name
@@ -265,26 +282,13 @@ def fetch_centre_slots(rdv_site_web, start_date, fetch_map: dict = None):
 
 
 def centre_iterator():
-    for centre in gouv_centre_iterator():
-        yield centre
-    for centre in ordoclic_centre_iterator():
-        yield centre
-    for centre in mapharma_centre_iterator():
-        yield centre
-    try:
-        center_path = 'data/output/doctolib-centers.json'
-        url = f"https://raw.githubusercontent.com/CovidTrackerFr/vitemadose/data-auto/{center_path}"
-        response = requests.get(url)
-        response.raise_for_status()
-        data = response.json()
-        file = open(center_path, 'w')
-        file.write(json.dumps(data, indent=2))
-        file.close()
-        logger.info(f"Found {len(data)} Doctolib centers (external scraper).")
-        for center in data:
-            yield center
-    except Exception as e:
-        logger.warning(f"Unable to scrape doctolib centers: {e}")
+    visited_centers_links = set()
+    for iterator in (ordoclic_centre_iterator, mapharma_centre_iterator,
+                     doctolib_center_iterator, gouv_centre_iterator):
+        for center in iterator():
+            if center["rdv_site_web"] not in visited_centers_links:
+                visited_centers_links.add(center["rdv_site_web"])
+                yield center
 
 
 def gouv_centre_iterator(outpath_format='data/output/{}.json'):
@@ -298,28 +302,32 @@ def gouv_centre_iterator(outpath_format='data/output/{}.json'):
 
     total = 0
 
-    centres_non_pris_en_compte = {"centres_fermes":{}, "centres_urls_vides":[]}
+    centres_non_pris_en_compte = {
+        "centres_fermes": {}, "centres_urls_vides": []}
 
     for row in csvreader:
 
         row["rdv_site_web"] = fix_scrap_urls(row["rdv_site_web"])
         if row["centre_fermeture"] == "t":
-            centres_non_pris_en_compte["centres_fermes"][row["gid"]] = row["rdv_site_web"]
+            centres_non_pris_en_compte["centres_fermes"][row["gid"]
+                                                         ] = row["rdv_site_web"]
 
         if len(row["rdv_site_web"]):
             yield row
         else:
-            centres_non_pris_en_compte["centres_urls_vides"].append(row["gid"])            
+            centres_non_pris_en_compte["centres_urls_vides"].append(row["gid"])
 
         total += 1
-    
+
     nb_fermes = len(centres_non_pris_en_compte["centres_fermes"])
     nb_urls_vides = len(centres_non_pris_en_compte["centres_urls_vides"])
 
-    logger.info(f"Il y a {nb_fermes} centres fermes dans le fichier gouv sur un total de {total}")
+    logger.info(
+        f"Il y a {nb_fermes} centres fermes dans le fichier gouv sur un total de {total}")
 
     nb_urls_vides = len(centres_non_pris_en_compte["centres_urls_vides"])
-    logger.info(f"Il y a {nb_urls_vides} centres avec une URL vide dans le fichier gouv sur un total de {total}")
+    logger.info(
+        f"Il y a {nb_urls_vides} centres avec une URL vide dans le fichier gouv sur un total de {total}")
 
     outpath = outpath_format.format("centres_non_pris_en_compte_gouv")
     with open(outpath, "w") as fichier:
